cleaned up the initial blockly tests
